pyproject.toml
[tool.poetry]
name = "ackack"
version = "0.1.0"
description = ""
authors = ["David Francos <opensource@davidfrancos.net>"]
readme = 'Readme.rst'

[tool.poetry.dependencies]
python = "^3.9"
weback-unofficial = "^0.3.5"
requests = "^2.27.1"
boto3 = "^1.20.31"
fastapi = "^0.71.0"
uvicorn = "^0.16.0"
PyQt5 = {version = "^5.15.6", optional = true, extras = ["game"]}
aiohttp = "^3.8.1"
rtsp = "^1.1.12"
loguru = "^0.6.0"
httpx = "^0.23.0"
websockets = "^10.3"
websocket-client = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
robot = 'ackack.robot:main'

[tool.poetry.extras]
game = ["PyQt5"]
